ShapeFromShading.jl

The following documents the functionality currently available in ShapeFormShading.jl.

Normal Integration:

ShapeFromShading.Frankot — Method

Frankot()

Defines the Frankot integrator which contians the Frankot-Chellappa method of integration. The Frankot-Chellappa method is a fast, reliable method for integration surface normals while enforcing integrability using Fourier methods.

Output

Frankot() returns a Frankot integrator which can then be called to run the Frankot-Chellappa method on a gradient field.

Details

Frankot-Chellappa method uses Fourier methods to attempt to solve the Poission equation $\nabla^2z = \partial_up + \partial_vq$. By taking the Fourier transform of both sides we get:

\[−(\omega^2_u + \omega^2_v)\hat{z}(\omega_u, \omega_v) = \imath \omega_u\hat{p} (\omega_u, \omega_v) + \imath \omega_v\hat{q}(\omega_u, \omega_v)\]

By rearanging the above equation we arive at an equation for $\hat{z}$;

\[\hat{z}(\omega_u, \omega_v) = \frac{\omega_u\hat{p}(\omega_u, \omega_v) + \omega_v\hat{q}(\omega_u, \omega_v)}{\imath(\omega^2_u + \omega^2_v)}\]

From which the final surface can be found by taking the inverse Fourier transform of $\hat{z}$.

Due to the way $(\omega_u, \omega_v)$ is defined the algorithm works best when the input dimentions are odd in length. To accomidate this the integrator will pad the edge of the inputs if they are even before running the algorithm. This padding will be removed before returning a value hence output size will be unaffected.

Parameters

Frankot integrator take no parameters.

Example

The following example demontraits the use of the Frankot integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Frankot() integrator
frankot = Frankot()

# Calculate the heightmap from the gradients
Z = frankot(p, q)

# Normalize to maximum of 1 (not necessary but makes displaying easier)
Z = Z./maximum(Z)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

R. T. Frankot and R. Chellappa, "A method for enforcing integrability in shape from shading algorithms," in IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 10, no. 4, pp. 439-451, July 1988. doi: 10.1109/34.3909

ShapeFromShading.Path — Method

Path()

Creates a Path() integrator which utilizes the average of two path integrals along varying paths. Each path integral reconstructs the surface with accumilating error along the path, hence averaging two different paths can minimize this error, although the method still suffers if the gradient field is not integrable at some points.

Output

Path() returns a Path integrator which can then be called to integrate a gradient field.

Details

Under the assumption that the surface normals are approximitly integrable everywhere ($\frac{\partial p}{\partial y}\approx\frac{\partial q}{\partial x}$), then surface can be reconstructed using the path integral defined as:

\[z(x,y)=\oint_c\left(\frac{\partial z}{\partial x},\frac{\partial z}{\partial y}\right)\cdot dl\]

Which can be broken into two integrals representing the value at each point on the surface as shown below for a path which integrates along the first column then along the row.

\[z(u,v)=\int_0^v\frac{\partial z}{\partial y}(0,y)dy + \int_0^u\frac{\partial z}{\partial x}(x,v)dx\]

The second path used in the algorithm is simply the transpose of the first, integrating along the first row then down the column represented mathmatically as:

\[z(u,v)=\int_0^u\frac{\partial z}{\partial x}(x,0)dx + \int_0^v\frac{\partial z}{\partial y}(u,y)dy\]

The algorithm can be writen, then discreatizes as shown below:

\[\begin{gathered} z(u,v)=\frac{1}{2}\left(\int_0^v\frac{\partial z}{\partial y}(0,y)dy + \int_0^u\frac{\partial z}{\partial x}(x,v)dx + \int_0^u\frac{\partial z}{\partial x}(x,0)dx + \int_0^v\frac{\partial z}{\partial y}(u,y)dy\right)\\ z(u,v)=\frac{1}{2}\left(\sum_{i=0}^vq(0,i) + \sum_{j=0}^up(j,v) + \sum_{j=0}^up(j,0) + \sum_{i=0}^vq(u,i)\right)\\ z(u,v)=\frac{1}{2}\left(\sum_{i=0}^v(q(0,i) + q(u,i)) + \sum_{j=0}^u(p(j,0) + p(j,v))\right)\\ \end{gathered}\]

It is important to note as mentioned above if there are non-integrable points in the normal field then artifacts can appear in the reconstruction. This is seen in the example below where the otherwise smooth sphere appears "spiky". This can be corrected post reconstruction by smoothing but idealy a different integrator should be used.

Arguments

Path integrator take no parameters.

Example

The following example demontraits the use of the Path integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Path() integrator
path = Path()

# Calculate the heightmap from the gradients
Z = path(p, q)

# Normalize to maximum of 1 (not necessary but makes displaying easier)
Z = Z./maximum(Z)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

D. Forsyth and J. Ponce, Computer vision: a modern approach. Upper Saddle River, N.J: Prentice Hall, 2003, pp. 84-86.

ShapeFromShading.SplitPath — Method

SplitPath()

Creates a SplitPath() integrator which utilizes the average of two path integrals along varying paths averaging the value at each step. Each path integral reconstructs the surface with accumilating error along the path, hence averaging two different paths at each step reduces the global error at the cost of local error, although the method still suffers if the gradient field is not integrable at some points it does less so the Path() from which it extends.

Output

SplitPath() returns a SplitPath integrator which can then be called to integrate a gradient field.

Details

Under the assumption that the surface normals are approximitly integrable everywhere ($\frac{\partial p}{\partial y}\approx\frac{\partial q}{\partial x}$), then surface can be reconstructed using the path integral defined as:

\[z(x,y)=\oint_c\left(\frac{\partial z}{\partial x},\frac{\partial z}{\partial y}\right)\cdot dl\]

By expanding on this principle and the discreate summation from Path() we can arrive at the discreate expresion for the value at each point, assuming all values prior to that point have been calculated, as follows:

\[z_{u,v} = \frac{1}{2}(z_{u-1,v}+p_{u-1,v}+z_{u,v-1}+q_{u,v-1})\]

As with other simular methods (see Horn()) care must be taken with regards to boundries which can be calculated, to a constant value $z(0,0)$ which is assumed to be the zero point, using:

\[\begin{gathered} z_{u,0} = z_{u-1,0}+p_{u-1,0}\\ z_{0,v} = z_{0,v-1}+q_{0,v-1} \end{gathered}\]

It is important to note as mentioned above if there are non-integrable points in the normal field then artifacts can appear in the reconstruction. These errors gradully average out but will lead to "streaks" appearing in the reconstruction. This is seen in the example below where the otherwise smooth sphere appears has ripple like structures pointing toward to top right corner. This can be corrected post reconstruction by smoothing but idealy a different integrator should be used. It is also interesting to note the parrallels between this method and the Horn and Brooks methed, with this mehtod being effectivly the forward component of Horn's method. As such this algorithm provided a usefull middle gorund between direct integration algorithms and iterative algorithms such as the Horn and Brooks mehtod.

Arguments

SplitPath integrator take no parameters.

Example

The following example demontraits the use of the Path integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Path() integrator
splitPath = SplitPath()

# Calculate the heightmap from the gradients
Z = splitPath(p, q)

# Normalize to maximum of 1 (not necessary but makes displaying easier)
Z = Z./maximum(Z)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

[1] D. Forsyth and J. Ponce, Computer vision: a modern approach. Upper Saddle River, N.J: Prentice Hall, 2003, pp. 84-86. [2] B. Horn and M. Brooks, "The variational approach to shape from shading", Computer Vision, Graphics, and Image Processing, vol. 33, no. 2, pp. 174-208, 1986. doi: 10.1016/0734-189x(86)90114-3

ShapeFromShading.Horn — Method

Horn()

Impliments the Horn and Brook's method of integrating surface normals. This algorithm offers an iterative solution to the Poisson equation descriping the surface providing good reconstructions under most conditions.

Output

Horn() returns a Horn integrator which can then be called to integrate a gradient field.

Details

The Horn and Brook's method attempts to solve the Poisson equation $\nabla^2z = \partial_up + \partial_vq$ by the discreatation given below:

\[z_{u+1,v}+z_{u,v+1}+z_{u-1,v}+z_{u,v-1}-4z_{u,v}=\frac{p_{u+1,v}-p_{u-1,v}}{2}+\frac{q_{u,v+1}-q_{u,v-1}}{2}\]

Which can be rearanged to give the iterative scheme provided by:

\[z_{u,v}^{k+1}= \frac{z_{u+1,v}^k + z_{u,v+1}^k + z_{u-1,v}^k + z_{u,v-1}^k}{4} - \frac{p_{u+1,v}-p_{u-1,v}}{8} - \frac{q_{u,v+1}-q_{u,v-1}}{8}\]

This scheme will always converge to a solution however the rate of convergance may depend upon the initial solution. This implimentation will initilize with a zero solution. Neumann boundry conditions are imposed at the edges where the scheme would otherwise go out of bounds.

Arguments

The function parameters are described in more detail below.

Max_iter

An Int which controls the numbe rof iterations the algorithm will run for. the range [0,1].

ϵ

A Real representing the distance between pixels. This will Control how tall the final reconstruction is relative the array grid.

Example

The following example demontraits the use of the Path integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Horn() integrator
horn = Horn(ϵ = 0.03, max_iter = 10000)

# Calculate the heightmap from the gradients
Z = horn(p, q)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

B. Horn and M. Brooks, "The variational approach to shape from shading", Computer Vision, Graphics, and Image Processing, vol. 33, no. 2, pp. 174-208, 1986. doi: 10.1016/0734-189x(86)90114-3

ShapeFromShading.Durou — Method

Durou()

Impliments the Durou and Courteille method of integrating surface normals. This algorithm offers an iterative solution to the Poisson equation descriping the surface extending Horn anf Brook's method by improving the boundry approximation and providing good reconstructions under most conditions.

Output

Durou() returns a Horn integrator which can then be called to integrate a gradient field.

Details

TheDurou and Courteille's method attempts to solve the Poisson equation $\nabla^2z = \partial_up + \partial_vq$ by the discreatation given below:

\[z_{u+1,v}+z_{u,v+1}-2z_{u,v}=\frac{p_{u+1,v}+p_{u,v}}{2}+\frac{q_{u,v+1}+q_{u,v}}{2}\]

Which can be rearanged to give the iterative scheme provided by:

\[z_{u,v}^{k+1}= \frac{z_{u+1,v}^k + z_{u,v+1}^k}{2} - \frac{p_{u+1,v}+p_{u,v}}{4} - \frac{q_{u,v+1}+q_{u,v}}{4}\]

This scheme will always converge to a solution however the rate of convergance may depend upon the initial solution. This implimentation will initilize with a zero solution. Natural boundry conditions are imposed at the edges using the condition ``\partialuz-p+\partialv-q=0. Although faster then the Horn and Brook's method and better at handeling boundries, it can generate a worse solution under some conditions.

Arguments

The function parameters are described in more detail below.

Max_iter

An Int which controls the numbe rof iterations the algorithm will run for. the range [0,1].

ϵ

A Real representing the distance between pixels. This will Control how tall the final reconstruction is relative the array grid.

Example

The following example demontraits the use of the Path integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Durou() integrator
durou = Durou(ϵ = 0.03, max_iter = 10000)

# Calculate the heightmap from the gradients
Z = durou(p, q)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

[1] Y. Quéau, J. Durou and J. Aujol, "Normal Integration: A Survey", Journal of Mathematical Imaging and Vision, vol. 60, no. 4, pp. 576-593, 2017. doi: 10.1007/s10851-017-0773-x