ShapeFromShading.jl

Documentation for ShapeFromShading.jl Work in progress mainly for testing at this point.

Missing docstring.

Missing docstring for Frankot(). Check Documenter's build log for details.

ShapeFromShading.Path — Method

Path()

Creates a Path() integrator which utilizes the average of two path integrals along varying paths. Each path integral reconstructs the surface with accumilating error along the path, hence averaging two different paths can minimize this error, although the method still suffers if the gradient field is not integrable at some points.

Output

Path() returns a Path integrator which can then be called to integrate a gradient field.

Details

Under the assumption that the surface normals are approximitly integrable everywhere ($\frac{\partial p}{\partial y}\approx\frac{\partial q}{\partial x}$), then surface can be reconstructed using the path integral defined as:

\[z(x,y)=\oint_c\left(\frac{\partial z}{\partial x},\frac{\partial z}{\partial y}\right)\cdot dl\]

Which can be broken into two integrals representing the value at each point on the surface as shown below for a path which integrates along the first column then along the row.

\[z(u,v)=\int_0^v\frac{\partial z}{\partial y}(0,y)dy + \int_0^u\frac{\partial z}{\partial x}(x,v)dx\]

The second path used in the algorithm is simply the transpose of the first integrating along the first row then down the column represented mathmatically as:

\[z(u,v)=\int_0^u\frac{\partial z}{\partial x}(x,0)dx + \int_0^v\frac{\partial z}{\partial y}(u,y)dy\]

The algorithm can be writen, then discreatizes as shown below:

\[\begin{gathered*} z(u,v)=\frac{1}{2}\left(\int_0^v\frac{\partial z}{\partial y}(0,y)dy + \int_0^u\frac{\partial z}{\partial x}(x,v)dx + \int_0^u\frac{\partial z}{\partial x}(x,0)dx + \int_0^v\frac{\partial z}{\partial y}(u,y)dy\right)\\ z(u,v)=\frac{1}{2}\left(\sum_{i=0}^vq(0,i) + \sum_{j=0}^up(j,v) + \sum_{j=0}^up(j,0) + \sum_{i=0}^vq(u,i)\right)\\ z(u,v)=\frac{1}{2}\left(\sum_{i=0}^v(q(0,i) + q(u,i)) + \sum_{j=0}^u(p(j,0) + p(j,v))\right)\\ \end{gathered*}\]

It is important to note as mentioned above if there are non-integrable points in the normal field then artifacts can appear in the reconstruction. This is seen in the example below where the otherwise smooth sphere appears "spiky". This can be corrected post reconstruction by smoothing but idealy a different integrator should be used.

Arguments

Path integrator take no parameters.

Example

The following example demontraits the use of the Path integrator.

using ShapeFromShading, Makie

# Generate synthetic gradients
p, q = synthetic_gradient(SynthSphere(), radius = 38, img_size = 151)

# Create a Path() integrator
path = Path()

# Calculate the heightmap from the gradients
Z = path(p, q)

# Normalize to maximum of 1 (not necessary but makes displaying easier)
Z = Z./maximum(Z)

# Display using Makie (Note: Makie can often take several minutes first time)
r = 0.0:0.1:4
surface(r, r, Z)

Reference

D. Forsyth and J. Ponce, Computer vision: a modern approach. Upper Saddle River, N.J: Prentice Hall, 2003, pp. 84-86.